Keep record of joined userIds in area.
Adds the server-assigned clientid to the area, allowing the server to exclusively broadcast to this area without checking each clients area, instead pulling a list of them from the area itself and sending exclusively to them.

Tests have been adjusted to account that area leaving and joining can now be identified by userid.